+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## v1.4.0
  - Cache instance list for five minutes before refreshing the list from AWS.
  - Show instance state in instance selection.